Some keyboard shortcuts added and language wise folder implementation done

+++ include/stop.php
@@ -1,8 +1,10 @@
 <?php
 session_start();
+if(isset($_SESSION['u_user'])){
 $pid=$_SESSION['pid'];
 /*echo $pid;
 ob_flush();
 flush();*/
 shell_exec("kill -9 ".$pid." ".($pid+1));
+}
 ?>
